Move `aspa_store` into separate file
